Move the definition of ErrorReporter to a separate header
